# backend/contracts/api/openapi.yaml
openapi: 3.1.0
info:
  title: Comfort Curators V0 API Contract Floor
  version: 0.1.0
  description: >-
    Minimum public and operational REST surface for the Comfort Curators V0
    modular monolith. Domain requirements and authorization rules in the
    protected product files remain normative.
servers:
  - url: http://localhost:8080
security:
  - bearerAuth: []
...
paths:
...
  /v1/superhost/runs:
    post:
      operationId: createSuperhostRun
      tags: [Automation]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/AgentRunCreate'}
      responses:
        '202': {$ref: '#/components/responses/Accepted'}
  /v1/jarvis/runs:
    post:
      operationId: createJarvisRunDeprecated
      deprecated: true
      description: >-
        Deprecated redirect alias preserved for one release, returning 308
        Permanent Redirect to /v1/superhost/runs. The client should resend the
        same method and body per HTTP semantics.
      tags: [Automation]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/AgentRunCreate'}
      responses:
        '308':
          description: Permanent redirect to /v1/superhost/runs
          headers:
            Location:
              schema: {type: string}
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
...
  responses:
...
    Accepted:
      description: Durable command accepted
      content:
        application/json:
          schema: {$ref: '#/components/schemas/Accepted'}
...
  schemas:
    Identifier:
      type: string
      minLength: 16
      maxLength: 64
      pattern: '^[A-Za-z0-9_-]+$'
...
    AgentRunCreate:
      type: object
      additionalProperties: false
      required: [idempotency_key, run_kind, tenant_id, property_id, trigger_type, trigger_id, correlation_id]
      properties:
        idempotency_key: {type: string, minLength: 8, maxLength: 128}
        run_kind: {type: string, enum: [superhost, hermes]}
        tenant_id: {$ref: '#/components/schemas/Identifier'}
        property_id: {$ref: '#/components/schemas/Identifier'}
        trigger_type: {type: string, maxLength: 64}
        trigger_id: {$ref: '#/components/schemas/Identifier'}
        correlation_id: {$ref: '#/components/schemas/Identifier'}
        purpose: {type: string, maxLength: 500}
...
    Accepted:
      type: object
      additionalProperties: false
      required: [request_id, status, resource_id]
      properties:
        request_id: {$ref: '#/components/schemas/Identifier'}
        status: {type: string, enum: [accepted, queued]}
        resource_id: {$ref: '#/components/schemas/Identifier'}
...